+++ src/core/component/functional/helpers.ts
@@ -81,20 +81,21 @@ export function inheritContext(
 
 	const
 		props = ctx.$props,
-		parentProps = parentCtx.$props;
-
-	const linkedFields = {};
+		parentProps = parentCtx.$props,
+		linkedFields = {};
 
 	Object.keys(parentProps).forEach((prop) => {
 		const linked = parentCtx.$syncLinkCache.get(prop);
 
-		if (linked != null) {
-			Object.values(linked).forEach((link) => {
-				if (link != null) {
-					linkedFields[link.path] = prop;
-				}
-			});
+		if (linked == null) {
+			return;
 		}
+
+		Object.values(linked).forEach((link) => {
+			if (link != null) {
+				linkedFields[link.path] = prop;
+			}
+		});
 	});
 
 	const fields = [