+++ src/core/component/interface/lc.ts
@@ -20,6 +20,7 @@ export type Hook =
 	'before:mounted' |
 	'mounted' |
 	'beforeUpdate' |
+	'before:updated' |
 	'updated' |
 	'activated' |
 	'deactivated' |