+++ src/entries/p-v4-components-demo.js
@@ -6,6 +6,8 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+import '../core/std';
+
 import '../core';
 import '../components/components/global/g-def';
 import '../components/pages/p-v4-components-demo';