+++ src/components/super/i-block/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.117 (2024-07-31)
+
+#### :house: Internal
+
+* Added a new `rootContent` layout wrapper block
+
 ## v4.0.0-beta.115.the-force-awakens (2024-07-26)
 
 #### :bug: Bug Fix